Store tuning parameters in structs (#731)
* Adding tuning params structs, doesn't build, rebase over this

* Working transition, now we fix up the tests

* Fix tests to match new scheme

Also, since blockSize and gridSize are 32-bit, actually make them
32-bit in our code.

* Use the getAttr() method

+++ mlir/lib/Conversion/MIOpenToGPU/MIOpenToGPU.cpp
@@ -142,8 +142,8 @@ void LowerMIOpenOpsToGPUPass::runOnOperation() {
     gpuModuleSymbolTable.insert(gpuFunc);
 
     // Set kernel attribute.
-    int64_t gridSize = 0;
-    int64_t blockSize = 0;
+    uint32_t gridSize = 0;
+    uint32_t blockSize = 0;
     gpuFunc->setAttr(gpu::GPUDialect::getKernelFuncAttrName(), b.getUnitAttr());
     if (auto attr = theFunc->getAttr("block_size")) {
       gpuFunc->setAttr("block_size", attr);
